Implement JAR parsing, organize javaparser-java classes and cleanup code
I was tasked with implementing a JAR reader into CodeCompass, which is able to read class files from JAR, decompile them, parse the decompiled code and persist them into database. It uses separate DecompileVisitor to recover compiled code as if it was original Java source,
then uses the existing original AstVisitor used for .java files.

The end result acts as a base, as it's only able to reverse engineer class structure with declared
fields and methods, however, local variable information gets lost during compilation and instructions can only be recovered as JVM assembly output. For the latter to be recovered, further bytecode analysis is needed. The JAR parsing was tested with a simple test JAR file and compilation
command like "javac -cp classinjar.jar:. Main.java".

During development, I also refactored and cleaned up the code, while trying to decouple responsibilities from Parser.java as much as possible, but it was not perfect as it was originally hardwired to be only used for parsing Java source files.

* Add new module for decompiling and parsing .class files in JAR archives
* Bundle Apache Commons Byte Code Engineering Library (BCEL, https://commons.apache.org/proper/commons-bcel) for class decompilation in JAR files
* Restructure existing classes into a package hierarchy, updating Makefile.am and persistence.xml accordingly
* Autoformat code with an IDE to remove hard tabs and use two spaces as soft tabs. Java-style braces are used.
* Because of JDK 7, changed unnecesseary template parameters in constructors to diamond operators ("List<String> ls = new ArrayList<>()")